# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DATN API",
        "description": "API for DATN project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/invoices/{invoiceId}/items": {
            "get": {
                "tags": [
                    "Invoice Items"
                ],
                "summary": "Danh sách hạng mục hóa đơn",
                "description": "Lấy danh sách tất cả hạng mục trong một hóa đơn",
                "operationId": "getInvoiceItems",
                "parameters": [
                    {
                        "name": "invoiceId",
                        "in": "path",
                        "description": "ID của hóa đơn",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách hạng mục",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/invoices/{invoiceId}/items/penalty": {
            "post": {
                "tags": [
                    "Invoice Items"
                ],
                "summary": "Thêm phí phát sinh vào hóa đơn",
                "description": "Thêm phí phát sinh, phí hư hại vào hóa đơn",
                "operationId": "addPenaltyItem",
                "parameters": [
                    {
                        "name": "invoiceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description",
                                    "quantity",
                                    "unit_price"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Phí hư hại thiết bị"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "unit_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500000
                                    },
                                    "item_type": {
                                        "type": "string",
                                        "enum": [
                                            "damage_fee",
                                            "other"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Phí được thêm thành công"
                    }
                }
            }
        },
        "/invoices/{invoiceId}/items/regular": {
            "get": {
                "tags": [
                    "Invoice Items"
                ],
                "summary": "Danh sách hạng mục thường trong hóa đơn",
                "description": "Lấy danh sách tất cả hạng mục thường (phòng, dịch vụ) trong một hóa đơn",
                "operationId": "getRegularItems",
                "parameters": [
                    {
                        "name": "invoiceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách hạng mục thường",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Invoice Items"
                ],
                "summary": "Thêm hạng mục vào hóa đơn",
                "description": "Thêm hạng mục (phòng, dịch vụ, khác) vào hóa đơn",
                "operationId": "addRegularItem",
                "parameters": [
                    {
                        "name": "invoiceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "item_name",
                                    "quantity",
                                    "unit_price",
                                    "item_type"
                                ],
                                "properties": {
                                    "item_name": {
                                        "type": "string",
                                        "example": "Phòng Deluxe"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "2 đêm"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "unit_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500000
                                    },
                                    "item_type": {
                                        "type": "string",
                                        "enum": [
                                            "room_charge",
                                            "service_charge",
                                            "damage_fee",
                                            "other"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Hạng mục được thêm thành công"
                    }
                }
            }
        },
        "/invoice-items/{id}": {
            "put": {
                "tags": [
                    "Invoice Items"
                ],
                "summary": "Cập nhật hạng mục hóa đơn",
                "description": "Cập nhật thông tin hạng mục hóa đơn",
                "operationId": "updateInvoiceItem",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "item_name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "unit_price": {
                                        "type": "number"
                                    },
                                    "item_type": {
                                        "type": "string",
                                        "enum": [
                                            "room_charge",
                                            "service_charge",
                                            "damage_fee",
                                            "other"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Invoice Items"
                ],
                "summary": "Xóa hạng mục hóa đơn",
                "description": "Xóa một hạng mục khỏi hóa đơn",
                "operationId": "deleteInvoiceItem",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with username and password to get the authentication token",
                "name": "Token based based security",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}